+++ resources/views/drivers.blade.php
@@ -37,7 +37,9 @@
                                     <td>{{$driver->sticker_number}}</td>
                                     <td>
                                         <a href="/operator/{{$driver->operator_id}}" class="btn btn-success custom-button-table"><i class="fa fa-info"></i> View Operator</a>
-                                        <a onClick="deleteDelete({{$driver->id}})" href="#" class="btn btn-danger custom-button-table"><i class="fa fa-trash"></i> Delete</a>
+                                        @if(Auth::user()->is_admin)
+                                            <a onClick="deleteDelete({{$driver->id}})" href="#" class="btn btn-danger custom-button-table"><i class="fa fa-trash"></i> Delete</a>
+                                        @endif
                                         </td>
                                 </tr>
                             @endforeach
