+++ resources/views/routes.blade.php
@@ -33,7 +33,9 @@
                                     <td>{{$route->rate_discounted}}</td>
                                     <td>
                                         <a href="/association/view/{{$route->assoc_id}}" class="btn btn-success custom-button-table"><i class="fa fa-info"></i> View Association</a>
-                                        <a href="#" onClick="deleteRouet({{$route->id}})" class="btn btn-danger custom-button-table"><i class="fa fa-trash"></i> Delete</a>
+                                        @if(Auth::user()->is_admin)
+                                            <a href="#" onClick="deleteRouet({{$route->id}})" class="btn btn-danger custom-button-table"><i class="fa fa-trash"></i> Delete</a>
+                                        @endif
                                     </td>
                                 </tr>
                             @endforeach
