+++ resources/views/view-association.blade.php
@@ -10,8 +10,10 @@
         <div class="col-md-12">
             <div class="card" style="margin-bottom: 31px;">
                <div class="card-header">
+                   @if(Auth::user()->is_admin)
                    <i class="fa fa-info-circle fa-lg"> {{$assoc->name_short}}</i> 
-                   <button onClick="updateAssoc()" type="submit" class="btn btn-primary to-right" style="margin-right: 10px;"><i class="fa fa-pencil-square-o fa-lg"></i> Update</button>
+                       <button onClick="updateAssoc()" type="submit" class="btn btn-primary to-right" style="margin-right: 10px;"><i class="fa fa-pencil-square-o fa-lg"></i> Update</button>
+                   @endif
                    <button  onClick="printSingleItem('association', {{$assoc->id}})"  class="btn btn-success to-right" style="margin-right: 10px;"><i class="fa fa-print fa-lg"></i> Print</button> 
                    <input id="csrf_token" type="hidden" value="{{ csrf_token() }}">
                </div>
@@ -74,9 +76,11 @@
             <div class="card">
                <div class="card-header">
                    <i class="fa fa-road fa-lg"></i> Routes
-                   <button data-toggle="modal" data-target="#route-modal" class="btn btn-primary to-right">
-                       <i class="fa fa-plus fa-lg"></i> Register New Route
-                   </button>
+                   @if(Auth::user()->is_admin)
+                       <button data-toggle="modal" data-target="#route-modal" class="btn btn-primary to-right">
+                           <i class="fa fa-plus fa-lg"></i> Register New Route
+                       </button>
+                   @endif
                    <button  onClick="printItems('route')" class="btn btn-success to-right" style="margin-right: 10px;">
                         <i class="fa fa-print fa-lg"></i> Print
                    </button> 
@@ -89,7 +93,9 @@
 				<th>Description</th>
 				<th>Rate</th>
 				<th>Rate w/ Discount</th>
-				<th>Option</th>
+                                @if(Auth::user()->is_admin)
+                                    <th>Option</th>
+				@endif
 			    </tr>
 			</thead>
 			<tbody>
@@ -99,14 +105,16 @@
                                 <td>{{$route->route}}</td>
                                 <td>{{$route->rate}}</td>
                                 <td>{{$route->rate_discounted}}</td>
-                                <td>
-                                    <button data-toggle="modal" data-id="{{$route->id}}" data-route="{{$route->route}}" data-rate="{{$route->rate}}" data-rate_discounted="{{$route->rate_discounted}}" data-target="#update-route-modal" class="btn btn-primary">
-                                        <i class="fa fa-pencil-square-o fa-lg"></i> Update  
-                                    </button>
-                                    <button onClick="deleteRoute({{$route->id}})" class="btn btn-danger">
-                                        <i class="fa fa-trash fa-lg"></i> Delete  
-                                    </button>
-                                </td> 
+                                @if(Auth::user()->is_admin)
+                                    <td>
+                                        <button data-toggle="modal" data-id="{{$route->id}}" data-route="{{$route->route}}" data-rate="{{$route->rate}}" data-rate_discounted="{{$route->rate_discounted}}" data-target="#update-route-modal" class="btn btn-primary">
+                                            <i class="fa fa-pencil-square-o fa-lg"></i> Update  
+                                        </button>
+                                        <button onClick="deleteRoute({{$route->id}})" class="btn btn-danger">
+                                            <i class="fa fa-trash fa-lg"></i> Delete  
+                                        </button>
+                                    </td> 
+                                @endif
                                 </tr> 
                             @endforeach
 			</tbody>
